Move json related code to JsonUtil
